+++ CHANGELOG.md
@@ -15,9 +15,13 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Changed
 
+- SC-4547 disable swagger docs in production
+
 ### Security
+
 - SC-4506 Secure Find User Route. Access user list by students is allowed only if they are eligible to create teams.   
 - SC-4506 Secure Get User Route. Read user details may only users with STUDENT_LIST or TEACHER_LIST permissions
+
 ### Removed